master  - implement VoteService delete operation  - improve AbstractVoteServiceTest - make methods for generating allowed and restricted dateTime  - make some TODOs

+++ src/main/java/com/github/ivansenchukov/topjavagraduation/model/Dish.java
@@ -1,12 +1,13 @@
 package com.github.ivansenchukov.topjavagraduation.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Dish extends AbstractBaseEntity {
 
     //<editor-fold desc="Fields">
     private String name;
-    private long price;
+    private BigDecimal price;
 
     private Restaurant restaurant;
     private LocalDate date;
@@ -22,14 +23,14 @@ public class Dish extends AbstractBaseEntity {
         this(prototype.getId(), prototype.getName(), prototype.getPrice(), prototype.getRestaurant(), prototype.getDate());
     }
 
-    public Dish(String name, long price, Restaurant restaurant, LocalDate date) {
+    public Dish(String name, BigDecimal price, Restaurant restaurant, LocalDate date) {
         this.name = name;
         this.price = price;
         this.restaurant = restaurant;
         this.date = date;
     }
 
-    public Dish(Integer id, String name, long price, Restaurant restaurant, LocalDate date) {
+    public Dish(Integer id, String name, BigDecimal price, Restaurant restaurant, LocalDate date) {
         super(id);
         this.name = name;
         this.price = price;
@@ -48,11 +49,11 @@ public class Dish extends AbstractBaseEntity {
         this.name = name;
     }
 
-    public long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
